Move status from an enum to a table
When building out functionality around inserting status updates, I ran into an issue with trying to get the endpoint to accept the enum value. After researching the issue I found this (https://github.com/postgraphql/postgraphql/issues/388), leading me to believe that enum values are currently buggy in PostgraphQL. I decided that moving status from an enumeration to a table would be a good solution, as the ability to customize statuses from within an admin interface may be desirable in the future.