Fix bug where updating crossing status wasn't updating apollo store (#161)
Updating the status of a crossing wasn't correctly updating the apollo store. I found that this was due to the move from react-apollo 1.x to react-apollo 2.x. Queries in the store have moved from store.data.ROOT_QUERY to store.data.data. The naming of the keys has also been updated. This is a quick bug fix to restore functionality, not a stand in for a refactor.